+++ src/components/UI/popup/types.ts
@@ -1,0 +1,4 @@
+export interface PopupPropsInterface {
+  children: React.ReactNode;
+  onClose: () => void;
+}