+++ src/components/layouts/sidebar/settings/index.tsx
@@ -1,0 +1,14 @@
+import SidebarBox from "../sidebarBox";
+
+const SidebarSettings = () => {
+  return (
+    <SidebarBox>
+      <div className='sidebar-settings'>
+        <div className='sidebar-settings__title'>Settings</div>
+        
+      </div>
+    </SidebarBox>
+  );
+};
+
+export default SidebarSettings;