+++ src/components/layouts/sidebar/sidebarBox/types.ts
@@ -1,0 +1,3 @@
+export interface SidebarBoxProps {
+  children: React.ReactNode;
+}