+++ src/components/layouts/sidebar/userInfo/changePassForm/types.ts
@@ -1,0 +1,14 @@
+export interface ChangePasswordFormProps {
+  onClose: () => void;
+  onSubmit: (values: ChangePasswordFormDataInterface) => void;
+}
+
+export interface ChangePasswordInputInterface {
+  password: string;
+  confirmPassword: string;
+}
+
+export interface ChangePasswordFormDataInterface {
+  password: string;
+  confirmPassword: string;
+}