+++ src/services/api/auth/login/index.ts
@@ -1,8 +1,7 @@
 import HttpClient from "../../../../helpers/axios";
 import { LoginApiResponseInterface, LoginParamsInterface } from "./types";
 
-const baseUrl = process.env.REACT_APP_BASE_URL;
-const URL = baseUrl + "/auth/login";
+const URL = "/auth/login";
 
 export async function login(params: LoginParamsInterface) {
   const response = await HttpClient.post<LoginApiResponseInterface>(