+++ src/services/api/offices/detail-office/index.ts
@@ -1,0 +1,9 @@
+import HttpClient from "../../../../helpers/axios";
+
+const URL = "/users/me/profile";
+
+
+export async function createOffice() {
+
+  }
+