+++ src/services/api/offices/detail-office/types.ts
@@ -1,0 +1,10 @@
+import { OfficeInterface } from "../../../../types/office";
+
+export interface OfficeApiResponseInterface {
+  office: OfficeInterface;
+  status?: string;
+  code?: number;
+  message?: string;
+  errors?: any[];
+}
+