+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Các thay đổi:
 
+## 26/03/2022
+
+- [feature] Thêm `InputDate`
+
+## 24/03/2022
+
+- [ui] thêm `sidebar`
+- [ui] thêm mục `user profile` ở `sidebar`
+- [ui] thêm mục `settings` ở `sidebar`
+- [ui] thêm mục `offices` ở `sidebar`
+
 ## 20/03/2022
 
 - [improve] setup toastify